+++ MatrixSDKTests/MatrixSDKTestsData.m
@@ -1045,6 +1045,7 @@ onUnrecognizedCertificateBlock:(MXHTTPClientOnUnrecognizedCertificate)onUnrecogn
         }
     }
     _retainedObjects = nil;
+    [MXRealmCryptoStore deleteAllStores];
 }
 
 @end